Capitalize POWER, Add documentation, write what true mean, Rename setMot
