+++ src/main/java/frc/robot/RobotContainer.java
@@ -12,7 +12,6 @@ import edu.wpi.first.wpilibj.GenericHID;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
-import frc.robot.subsystems.intake.Intake;
 import frc.robot.subsystems.intake.commands.Commandy;
 
 
@@ -26,7 +25,7 @@ public class RobotContainer {
   public XboxController Xbox = new XboxController(1);
   public JoystickButton a = new JoystickButton(Xbox, XboxController.Button.kA.value);
   public JoystickButton b = new JoystickButton(Xbox, XboxController.Button.kB.value);
-  private Intake intake;
+  private final Intake intake= new Intake();
 
   // The robot's subsystems and commands are defined here...
 
@@ -46,7 +45,7 @@ public class RobotContainer {
    * {@link edu.wpi.first.wpilibj2.command.button.JoystickButton}.
    */
   private void configureButtonBindings() {
-    a.whileHeld(new Commandy(intake, Constants.Intake.power));
+    a.whileHeld(new Commandy(intake, Constants.Intake.POWER));
   }
